支持txt导入联赛数据 自动分析持续时间，不活跃人数 自动提取帮众信息和帮会信息

+++ qnyh/youxi-collect/src/main/java/org/lingxi/youxi/collect/beans/OutputAmount.java
@@ -11,7 +11,7 @@ public class OutputAmount {
 	
 	private String name;//名称
 	private String profession;//职业
-	private Integer amount;//输出量
+	private Long amount;//输出量
 	/**
 	 * @return the name
 	 */
@@ -39,15 +39,16 @@ public class OutputAmount {
 	/**
 	 * @return the amount
 	 */
-	public Integer getAmount() {
+	public Long getAmount() {
 		return amount;
 	}
 	/**
 	 * @param amount the amount to set
 	 */
-	public void setAmount(Integer amount) {
+	public void setAmount(Long amount) {
 		this.amount = amount;
 	}
+
